renaming with Network naming
