--- common/src/androidMain/kotlin/me/kevinschildhorn/common/connection/data/File.kt
@@ -1,5 +1,0 @@
-package me.kevinschildhorn.common.connection.data
-
-import org.apache.commons.net.ftp.FTPFile
-
-actual typealias File = FTPFile
