+++ common/src/androidMain/kotlin/me/kevinschildhorn/common/network/data/NetworkFile.kt
@@ -1,0 +1,5 @@
+package me.kevinschildhorn.common.network.data
+
+import org.apache.commons.net.ftp.FTPFile
+
+actual typealias NetworkFile = FTPFile
