--- common/src/commonMain/kotlin/me/kevinschildhorn/common/connection/data/Directory.kt
@@ -1,3 +1,0 @@
-package me.kevinschildhorn.common.connection.data
-
-expect class Directory
