+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/network/DirectoryInformation.kt
@@ -1,0 +1,10 @@
+package me.kevinschildhorn.common.network
+
+import me.kevinschildhorn.common.network.data.NetworkDirectory
+import me.kevinschildhorn.common.network.data.NetworkFile
+
+data class DirectoryInformation(
+    val path: String,
+    val directories: List<NetworkDirectory>,
+    val files: List<NetworkFile>
+)
