+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/network/LoginInformation.kt
@@ -1,4 +1,4 @@
-package me.kevinschildhorn.common.connection
+package me.kevinschildhorn.common.network
 
 
 data class LoginInformation(
