+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/network/data/NetworkFile.kt
@@ -1,0 +1,3 @@
+package me.kevinschildhorn.common.network.data
+
+expect class NetworkFile
