--- common/src/iosMain/kotlin/me/kevinschildhorn/common/connection/data/File.kt
@@ -1,3 +1,0 @@
-package me.kevinschildhorn.common.connection.data
-
-actual typealias File = String
