+++ common/src/iosMain/kotlin/me/kevinschildhorn/common/network/ftps/FTPSClient.kt
@@ -1,6 +1,6 @@
-package me.kevinschildhorn.common.connection.ftps
+package me.kevinschildhorn.common.network.ftps
 
-import me.kevinschildhorn.common.connection.DirectoryInformation
+import me.kevinschildhorn.common.network.DirectoryInformation
 
 actual object FTPSClient {
     actual val isConnected: Boolean = false
